refactor: optimize import export hub for octokit api
