+++ src/db/controllers/index.ts
@@ -1,5 +1,10 @@
+import installationController from "./installation.controller.js";
+import userController from "./user.controller.js";
+
 /** ***************** **
  * ******************* *
  * Export Controllers **
  * ******************* *
  ** ***************** **/
+
+export { installationController, userController };
